pyproject.toml
[tool.poetry]
name = "order-automation"
version = "0.1.0"
description = ""
authors = ["Yalchin403 <yalchin.mammadli.std@bhos.edu.az>"]
readme = "README.md"
packages = [{include = "order_automation"}]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
amqp = "5.1.1"
asgiref = "3.5.2"
billiard = "3.6.4.0"
celery = "5.2.6"
click = "8.1.3"
click-didyoumean = "0.3.0"
click-plugins = "1.1.1"
click-repl = "0.2.0"
colorama = "0.4.5"
django = "4.1.2"
django-celery-results = "2.4.0"
django-jazzmin = "2.5.0"
kombu = "5.2.4"
packaging = "21.3"
pillow = "9.1.1"
prompt-toolkit = "3.0.29"
psycopg2-binary = "2.9.4"
pyotp = "2.6.0"
pyparsing = "3.0.9"
python-dotenv = "0.20.0"
pytz = "2022.1"
redis = "4.1.1"
six = "1.16.0"
sqlparse = "0.4.2"
tzdata = "2022.1"
vine = "5.0.0"
wcwidth = "0.2.5"
whitenoise = "6.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
